Move S3Config into the new storage lib
