+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessageWriter.scala
@@ -7,7 +7,6 @@ import com.amazonaws.services.sns.AmazonSNS
 import com.google.inject.Inject
 import com.twitter.inject.Logging
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
-import uk.ac.wellcome.models.aws.S3Config
 
 import io.circe.Encoder
 import uk.ac.wellcome.messaging.sns.SNSWriter
