+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/s3/S3Config.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.aws
+package uk.ac.wellcome.storage.s3
 
 case class S3Config(
   bucketName: String,
