+++ sierra_adapter/sierra_item_merger/src/test/scala/uk/ac/wellcome/platform/sierra_item_merger/services/SierraItemMergerUpdaterServiceTest.scala
@@ -8,13 +8,13 @@ import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{Assertion, FunSpec}
 import uk.ac.wellcome.messaging.test.fixtures.SQS
 import uk.ac.wellcome.metrics.MetricsSender
-import uk.ac.wellcome.models.aws.S3Config
 import uk.ac.wellcome.models.transformable.SierraTransformable
 import uk.ac.wellcome.models.{SourceMetadata, Sourced}
 import uk.ac.wellcome.platform.sierra_item_merger.utils.SierraItemMergerTestUtil
 import uk.ac.wellcome.storage.dynamo._
 import uk.ac.wellcome.storage.s3.{
   KeyPrefixGenerator,
+  S3Config,
   S3ObjectStore,
   SourcedKeyPrefixGenerator
 }
